+++ src/main/java/org/betonquest/betonquest/compatibility/jobsreborn/objective/PaymentObjective.java
@@ -9,6 +9,8 @@ import org.betonquest.betonquest.api.instruction.Instruction;
 import org.betonquest.betonquest.api.instruction.variable.Variable;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
+import org.betonquest.betonquest.api.quest.objective.ObjectiveData;
+import org.betonquest.betonquest.api.quest.objective.ObjectiveDataFactory;
 import org.betonquest.betonquest.quest.event.IngameNotificationSender;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
