+++ src/main/java/org/betonquest/betonquest/quest/objective/stage/StageObjective.java
@@ -6,6 +6,8 @@ import org.betonquest.betonquest.api.Objective;
 import org.betonquest.betonquest.api.instruction.Instruction;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
+import org.betonquest.betonquest.api.quest.objective.ObjectiveData;
+import org.betonquest.betonquest.api.quest.objective.ObjectiveDataFactory;
 import org.betonquest.betonquest.api.quest.objective.ObjectiveID;
 
 import java.util.List;
@@ -156,7 +158,7 @@ public class StageObjective extends Objective {
     }
 
     /**
-     * {@link org.betonquest.betonquest.api.Objective.ObjectiveData} for {@link StageObjective}.
+     * {@link ObjectiveData} for {@link StageObjective}.
      */
     public static class StageData extends ObjectiveData {
         /**
